requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
defusedxml==0.7.1
ffmpeg-python==0.2.0
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
future==1.0.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
openai-whisper==20250625
pillow==12.0.0
psutil==7.1.3
pydub==0.25.1
regex==2025.10.23
requests==2.32.5
six==1.17.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
youtube-transcript-api==1.2.3
yt-dlp==2025.10.22
